new component SalleResa + disabled all resa date ok
