+++ .gitignore
@@ -1,5 +1,8 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
+# general
+.vscode
+
 # dependencies
 /node_modules
 /.pnp
