+++ src/pages/Salles/Salles.js
@@ -6,19 +6,18 @@ import useAxios from '../../hooks/useAxios/useAxios';
 
 import './Salles.css'
 import 'react-calendar/dist/Calendar.css';
+import SalleResa from '../../components/SalleResa/SalleResa';
 
 const Salles = () => {
     const keys = ['nom', 'description', 'capacite', 'prix'];
 
     const {response} = useAxios('get', 'http://localhost:3001/salles/');
-
-    // const {response} = useAxios('get', 'http://localhost:3001/users/all/reservations/');
-
-    console.log(response);
-
+    const {response: allResa} = useAxios('get', 'http://localhost:3001/users/all/reservations/');
+    
     const [ headers, setHeaders ] = useState();
     const [ items, setItems ] = useState();
-    const [ idSalle, setIdSalle ] = useState();
+    const [ idSalle, setIdSalle ] = useState(5);
+    
     
 
     const addData = (e, response) => {
@@ -27,7 +26,7 @@ const Salles = () => {
             if (salle.nom === e.target.textContent || salle.nom === e.target.parentNode.children[1].textContent) {
                 setHeaders([ 'Nom', 'Description', 'Capacité', 'Prix (€)']);
                 setIdSalle(salle.id);
-                console.log(idSalle);
+                // console.log(idSalle);
                 setItems([salle]);
             }
         });
@@ -44,25 +43,15 @@ const Salles = () => {
         addData(e, response);
     }
 
-    const selectDay = (e) => {
-        let jour = new Date(e).toLocaleString().split(",")[0].split("/");
-        jour = `${jour[2]}-${jour[1]}-${jour[0]}`;
-        console.log(jour, idSalle);
-    }
-
     return (  
         <div className="salles">
             <h1>Salles mise à disposition des ligues</h1>
             <PlanSalles selectSalle={selectSalle}/>
-            <div className="list_resa">
-                { items &&
-                    <Calendar onClickDay={selectDay} tileDisabled={({activeStartDate, date, view }) => {
-                        console.log(date)
-                        if(date.getDay() === 0 || date.getDay() === 6) return true;
-
-                    }} /> }
+            {items &&
+                <div className="list_resa">
+                <SalleResa idSalle={idSalle} dataResa={allResa} />
                 <ItemList name='salles' keys={keys} headers={headers} data={items}/>
-            </div>
+            </div>}
         </div>
     );
 }
